requirements.txt
fastapi
uvicorn
sqlalchemy
pyodbc
pandas
scikit-learn
streamlit
python-dotenv
openai
google-generativeai
requests
pydantic